pyproject.toml
[project]
name = "gemini-react-agent"
version = "0.1.0"
description = "A ReAct (Reason and Act) AI agent using Google Gemini"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-genai>=0.12.0",
    "python-dotenv>=1.0.1",
    "yfinance>=0.2.54",
]